Remove the Groovydoc plugin use
This plugin causes a resolution error failure when releasing.
We switch to Gradle's groovydoc task instead, which shouldn't
be a problem because we have almost no Groovy classes in the
public API and it seems that Groovy docs are not published
as part of the documentation we build in `docs`.