refactor: Move forcing of groovy-xml to buildSrc
When using Gradle 7.6 together with the Grails Gradle Plugin and groovy files in src/*/groovy there is a clash between the version bundled with Gradle (3.0.13) and the version loaded by the Grails Gradle Plugin (3.0.11). This forces the version to be compatible with Gradle for compiling groovy files in buildSrc.

This was previously included in the application's build.gradle leading to unnecessary clutter.