Include Geb/Selenium dependencies for MacOS Arch64 (#385)
* fix: Enable inclusion of Geb/Selenium deps for Macos Arch64

This was already enabled in another part of the code in
ad660a56e969855d7792e7ffdfed26fc58b2759c

It looks like this was missed.

* test: Temporarily disable integrationTest task in create-app specs

Due to an issue with the integrationTest task during the create-app tests, the task is being temporarily disabled. The tests fail with `org.openqa.selenium.SessionNotCreatedException`, which seems to be related to the inclusion of Geb/Selenium dependencies for `OperatingSystem.MACOS_ARCH64`.

The apps in the create-app tests are created with `OperatingSystem.MACOS_ARCH64` as the operating system parameter, but will fail also for other operating systems.

This issue was likely masked before commit 8675723e62df6d136d7af48d5c75d7728cbef871, when the Geb/Selenium dependencies were not included for MACOS_ARCH64. With the dependencies now included, the integrationTest task fails.